Define rLens in terms of the more general rLens'.
Provide rLens' to produce lenses of record fields of an arbitrary
functor type. rLens is now simply a specialization of rLens' from Rec
to PlainRec.

The new rLens' obviates the need for a helper function that had been
added to the Intro doctest.

Fix cabal test integration with the doctests. Previously, cabal test
did not actually run the tests because the main action in Intro.lhs
didn't do anything. Now it explicitly calls doctest.

Bump the package version.