Implement "lazy" records using type families
Because `Rec` is a GADT, it cannot be used inside a lazy (~) pattern. Therefore certain recursive programs which require laziness cannot be written using Rec. This patch adds a parallel implementation, `LazyRec`, using type families to avoid that shortcoming.

The downside is that recursive functions are less convenient to write. You can't just pattern match on RNil/(:&), because RNilL and (:&~) are constructors of different types. You have to use type classes like `ToRec` and `FromRec`.