Remove the f-parameter from Rec
This simplifies the interface, but it does result in some trouble that
I'd prefer to avoid.

We get around having that parameter by just composing it with the el
parameter when needed. This has the added benefit, then, of records
being truly raw: no Identity newtype needed.

Problem is that GHC doesn't know that (Id . el) = el. This results in
the necessity of unsafeCoerce (or equivalent) even in consuming code. I
don't like that so much...