Equality definition benchmark
A quick demonstration that GHC is better able to optimize recursive type
class instances than recursive functions. The type class induction is 5x
faster (on some computers) than the recursive function for a simple
record.

I'm adding this benchmark now is it reflects a design decision Vinyl has
gone back and forth on over the years. Its presence can serve as a
reminder or past experiments, and a small framework for verifying
performance expectations as GHC evolves.