Added the RecMapMethod type class
It is a bit complex, but should provide good usability. It lets you
use instances for field types without considering the functor, even if
the "functor" is ElType with its phantom tag. As well as the class
method, `rmapMethod`, there are definitions for records with a
`Functor` context, `rmapMethodF`, and for `FieldRec`s, `mapFields`.

- Added more instances for ElField

- Removed some #if-guarded code for compiler versions (< 8) we no
  longer support.

Addresses #113