Add the RecPointed class
The use case here is slightly different than for RecMapMethod. In the
latter, we want to use instances on field indexes to manipulate those
fields. In the case of a `pointed` method, we want to construct values
from whole cloth, and for that the interpretation functor is
particularly relevant.

For example, a default value of an `ElField '("age", Int)` really
needs the interpretation functor. Likewise, for more generally useful
functors like `Maybe`.