XRec API changes
XRec is a type synonym: `type XRec f = Rec (XData f)`, and `HKD` is an
associated type family of the `IsoHKD` type class.

This is better in that `HKD` is now an open type family. It is worse
if every other interpretation functor uses the default instance. In
that case, the need to write `instance IsoHKD MyType where` for every
`MyType` is a nuisance we avoided with the closed type family's
catch-all case.